LayoutTests/fast/canvas/webgl/gl-pixelstorei-expected.txt


There should be 5 red triangles on 5 black squares above

This test checks that drawImage and readPixels are not effected by gl.Pixelstorei(gl.PACK_ALIGNMENT) and visa versa

PASS drawing is correct in 3d context
PASS drawing is correct in 2d context
PASS PACK_ALIGNMENT is 1
PASS drawing is correct in 3d context
PASS drawing is correct in 2d context
PASS PACK_ALIGNMENT is 2
PASS drawing is correct in 3d context
PASS drawing is correct in 2d context
PASS PACK_ALIGNMENT is 4
PASS drawing is correct in 3d context
PASS drawing is correct in 2d context
PASS PACK_ALIGNMENT is 8
PASS successfullyParsed is true

TEST COMPLETE
